Move keyboard shortcuts from index.ts to workspaces feature
